Plano aee - add parametro codigo turma na consulta do plano e da versao
