+++ src/SME.SGP.WebClient/src/paginas/Relatorios/AEE/Plano/Cadastro/Componentes/SecaoVersaoPlano/montarDadosPorSecaoVersao.js
@@ -1,29 +1,35 @@
 import PropTypes from 'prop-types';
-import React, { useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import QuestionarioDinamico from '~/componentes-sgp/QuestionarioDinamico/questionarioDinamico';
 import ServicoPlanoAEE from '~/servicos/Paginas/Relatorios/AEE/ServicoPlanoAEE';
 
 const MontarDadosPorSecaoVersao = props => {
-  const { versao, dados } = props;
+  const { versao, dados, questionarioId, exibir, turmaCodigo } = props;
   const [dadosQuestionarioAtual, setDadosQuestionarioAtual] = useState([]);
 
   const dadosCollapseLocalizarEstudante = useSelector(
     store => store.collapseLocalizarEstudante.dadosCollapseLocalizarEstudante
   );
 
-  const obterDadosPorVersaoId = async () => {
+  const obterDadosPorVersaoId = useCallback(async () => {
     if (versao > 0) {
-      const resposta = await ServicoPlanoAEE.obterVersaoPlanoPorId(versao);
+      const resposta = await ServicoPlanoAEE.obterVersaoPlanoPorId(
+        versao,
+        questionarioId,
+        turmaCodigo
+      );
       if (resposta?.data) {
         setDadosQuestionarioAtual(resposta?.data);
       }
     }
-  };
+  }, [questionarioId, versao, turmaCodigo]);
 
   useEffect(() => {
-    obterDadosPorVersaoId();
-  }, []);
+    if (exibir && !dadosQuestionarioAtual?.length) {
+      obterDadosPorVersaoId();
+    }
+  }, [exibir, dadosQuestionarioAtual, obterDadosPorVersaoId]);
 
   return dadosQuestionarioAtual?.length ? (
     <QuestionarioDinamico
@@ -44,11 +50,17 @@ const MontarDadosPorSecaoVersao = props => {
 MontarDadosPorSecaoVersao.propTypes = {
   dados: PropTypes.oneOfType([PropTypes.object]),
   versao: PropTypes.oneOfType([PropTypes.number]),
+  questionarioId: PropTypes.oneOfType([PropTypes.any]),
+  exibir: PropTypes.bool,
+  turmaCodigo: PropTypes.oneOfType([PropTypes.any]),
 };
 
 MontarDadosPorSecaoVersao.defaultProps = {
   versao: 0,
   dados: {},
+  questionarioId: '',
+  exibir: false,
+  turmaCodigo: '',
 };
 
 export default MontarDadosPorSecaoVersao;
